feat: the live conformance suite is 42/42 green on all five SDKs
Both remaining xfails fixed, in the engine, and both notes had the mechanism
wrong in a way worth recording.

STREAM_XACK now accepts the id XADD produced
  XADD returns "<ms>-<seq>" as one string; XACK took the two halves as separate
  integers. The two ends of the same API did not compose, so every SDK split
  the string itself and the consumer-group case was xfail in all five. Splitting
  it per client is five chances to disagree about a separator, so the engine
  accepts (stream, group, '<ms>-<seq>') alongside the four-argument form.

  This needed the wire parameter-type table to become arity-aware: it is keyed
  by function name and position, so the three-argument form had its id typed
  INT8 from the four-argument signature and could not bind. A table that is
  position-keyed but not arity-keyed is invisible until one function has two
  shapes, and this is the first that does.

KV_ZRANGE negative indices
  The note said it "reads -1 literally". It does not, and literal would have
  been better. The store took `usize`, so -1 became usize::MAX, and then
  `stop.saturating_sub(start) + 1` OVERFLOWED to 0 — `.take(0)` returned an
  empty vector. Reading -1 literally returns something; this returned nothing,
  on `zrange(key, 0, -1)`, the most idiomatic call in the surface.

  ZRANGE and ZREVRANGE now take signed indices through a resolver shared with
  LRANGE, so the three cannot drift again. LRANGE having been fixed years
  earlier while ZRANGE was missed is exactly how this survived: two functions,
  one semantic, two implementations.

  The RESP surface was NOT affected — it resolves negatives before calling the
  store, which is why a Redis client got the right answer and a SQL client did
  not.

Every SDK updated to the new xack signature (id as one string), and four unit
tests that asserted the old four-argument SQL were updated — they were pinning
the shape that caused the problem.

Live: python 42 · go 42 · typescript 42 · rust 42 · elixir 42.
0 fail · 0 xfail · 0 unsupported · 0 known-drift. First fully green run.
nucleus lib 4235 · go green · python 479 · elixir 495 · typescript 306.

+++ nucleus/src/embedded.rs
@@ -790,16 +790,16 @@ impl KvHandle<'_> {
     pub fn col_zrange(
         &self,
         key: &str,
-        start: usize,
-        stop: usize,
+        start: i64,
+        stop: i64,
     ) -> Result<Vec<crate::kv::SortedSetEntry>, crate::kv::collections::WrongTypeError> {
         self.store.col_zrange(key, start, stop)
     }
     pub fn col_zrevrange(
         &self,
         key: &str,
-        start: usize,
-        stop: usize,
+        start: i64,
+        stop: i64,
     ) -> Result<Vec<crate::kv::SortedSetEntry>, crate::kv::collections::WrongTypeError> {
         self.store.col_zrevrange(key, start, stop)
     }
